src/contracts/requirements.txt
Flask
Flask-SQLAlchemy
pulsar-client
pulsar-client[avro]
psycopg2-binary
PyDispatcher
requests